backend_p34/requirements.txt
#Install them locally/server: pip install -r requirements.txt
#Transfer all the backend files to CVM via: scp -r ~/Desktop/backend ubuntu@134.175.220.232:/home/ubuntu/
fastapi
uvicorn
pydantic
requests
python-dotenv
#logging  # Built-in, but for completeness
tcvectordb
psycopg2-binary
sentence-transformers
sqlalchemy
websockets
pyjwt
alembic
slowapi  # Optional for rate limiting